use Supports class for util classes
