--- maintools/src/main/java/org/huberb/h2tools/jdbc/ConsumerThrowingSQLException.java
@@ -1,58 +1,0 @@
-/*
- * Copyright 2022 berni3.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.huberb.h2tools.jdbc;
-
-import java.sql.SQLException;
-import java.util.Objects;
-
-/**
- * Define a consumer for jdbc operations.
- *
- * @author berni3
- * @param <T>
- */
-@FunctionalInterface
-public interface ConsumerThrowingSQLException<T> {
-
-    /**
-     * Performs this operation on the given argument.
-     *
-     * @param t the input argument
-     * @throws java.sql.SQLException
-     */
-    void accept(T t) throws SQLException;
-
-    /**
-     * Returns a composed {@code Consumer} that performs, in sequence, this
-     * operation followed by the {@code after} operation. If performing either
-     * operation throws an exception, it is relayed to the caller of the
-     * composed operation. If performing this operation throws an exception, the
-     * {@code after} operation will not be performed.
-     *
-     * @param after the operation to perform after this operation
-     * @return a composed {@code Consumer} that performs in sequence this
-     * operation followed by the {@code after} operation
-     * @throws NullPointerException if {@code after} is null
-     */
-    default ConsumerThrowingSQLException<T> andThen(ConsumerThrowingSQLException<? super T> after) {
-        Objects.requireNonNull(after);
-        return (T t) -> {
-            accept(t);
-            after.accept(t);
-        };
-    }
-
-}
