+++ maintools/src/test/java/org/huberb/h2tools/jdbc/JdbcSqlFTest.java
@@ -24,7 +24,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.h2.jdbcx.JdbcConnectionPool;
-import org.huberb.h2tools.jdbc.JdbcSqlF.Holder;
+import org.huberb.h2tools.jdbc.Supports.ConsumerThrowingSQLException;
+import org.huberb.h2tools.jdbc.Supports.Holder;
 import org.junit.jupiter.api.AfterEach;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
