+++ maintools/src/test/java/org/huberb/h2tools/jdbc/JdbcSqlTest.java
@@ -27,6 +27,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 import javax.sql.DataSource;
+import org.huberb.h2tools.jdbc.Supports.ConsumerThrowingSQLException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
